Rebuild
Most of the core classes have been slightly or completely changed, in
order to accommodate the new external character creator, and just
because it was due to be overhauled. It's also more robust, easier to
understand (I think) and prettier.

A few things aren't completely redone, though. Collisions in particular
haven't been re-implemented yet, since i'd like to change the way they
behave and make them more efficient. That's next on the list. :/