Hitstun, combo tracker and wallbounce.
Added-
>Hitstun, a period after each attack where the 'hurt' animation will be
played, and no player input will be allowed.
>A simple combotracker. If more than one attack lands in quick
succession, the combotracker will start. The 'tracker will display the
number of succesful attacks in this period, and the combined damage of
the combo. So long as the attacks are all landed while the target is in
hitstun, the display will be red, but if the combo continues after the
target has left hitstun (i.e. they had a chance to break out of the
combo) the display will be blue.
>Wallbouncing. Doesn't really work yet. Well, it works, but it's too
'bouncy.' Characters feel like they're made of rubber and there is no
sensation of momentum or impact.